# Conflicts:
#	.gradle/buildOutputCleanup/outputFiles.bin
#	src/main/java/com/contact/controller/WebController.java
#	src/main/java/com/contact/entity/Contact.java
#	src/main/resources/META-INF/persistence.xml
#	src/main/resources/templates/add.html
#	src/main/resources/templates/contact.html
#	src/main/resources/templates/home.html

+++ src/main/java/com/contact/entity/Contact.java
@@ -10,8 +10,7 @@ import java.util.List;
 public class Contact{
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private int id;
-
+    private long id;
     private String name;
     private String trymail;
 
@@ -25,13 +24,14 @@ public class Contact{
         this.mail = new ArrayList<String>();
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
     public void setName(String name) {
         this.name = name;
     }
+
     public String getName() {
         return name;
     }
@@ -45,6 +45,5 @@ public class Contact{
 
     @ManyToMany
     private Collection<Address> addresses;
-
 
 }
